Better abstract test fixtures and their tested diagnostics
