--- GenericsAnalyzer/GenericsAnalyzer.Test/IAnalyzerTestFixture.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-
-namespace GenericsAnalyzer.Test
-{
-    public interface IAnalyzerTestFixture
-    {
-        DiagnosticDescriptor TestedDiagnosticRule { get; }
-    }
-}
