+++ GenericsAnalyzer/GenericsAnalyzer.Test/PermittedTypeArguments/CodeFixes/GA0002_CodeFixTests.cs
@@ -1,13 +1,10 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace GenericsAnalyzer.Test.PermittedTypeArguments.CodeFixes
 {
     [TestClass]
     public class GA0002_CodeFixTests : DuplicateAttributeArgumentRemoverCodeFixTests
     {
-        public override DiagnosticDescriptor TestedDiagnosticRule => DiagnosticDescriptors.GA0002_Rule;
-
         // TODO: Find a way to test applying the code fix on a specified diagnostic
         [TestMethod]
         public void ConflictingConstraintsCodeFix()
