+++ GenericsAnalyzer/GenericsAnalyzer.Test/PermittedTypeArguments/CodeFixes/GA0011_CodeFixTests.cs
@@ -1,13 +1,10 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace GenericsAnalyzer.Test.PermittedTypeArguments.CodeFixes
 {
     [TestClass]
     public class GA0011_CodeFixTests : RedundantAttributeArgumentRemoverCodeFixTests
     {
-        public override DiagnosticDescriptor TestedDiagnosticRule => DiagnosticDescriptors.GA0011_Rule;
-
         [TestMethod]
         public void RedundantlyProhibitedTypeWithCodeFix()
         {
